+++ AppKit/NSColorPicker.subproj/NSColorPickerWheelView.m
@@ -22,9 +22,9 @@
 	NSRect bounds = [self bounds];
 	NSPoint location = NSMakePoint(NSMidX(bounds), NSMidY(bounds));
 	
-	float colorWheelRadius = .5 * MIN(NSWidth(bounds), NSHeight(bounds));
+	CGFloat colorWheelRadius = .5 * MIN(NSWidth(bounds), NSHeight(bounds));
 	
-	float angle = M_PI * _hueValue/ 180.f;
+	CGFloat angle = M_PI * _hueValue/ 180.f;
 	
 	location.x -= colorWheelRadius * sin(angle) * _saturationValue/100.f;
 	location.y += colorWheelRadius * cos(angle) * _saturationValue/100.f;
@@ -121,18 +121,18 @@
 	NSPoint location = [self convertPoint:[event locationInWindow] fromView:nil];
 	
 	NSRect bounds = [self bounds];
-	float colorWheelRadius = .5 * MIN(NSWidth(bounds), NSHeight(bounds));
+	CGFloat colorWheelRadius = .5 * MIN(NSWidth(bounds), NSHeight(bounds));
 	
 	NSPoint center = NSMakePoint(NSMidX(bounds), NSMidY(bounds));
 	location.x -= center.x;
 	location.y -= center.y;
 	
-	float relativeDistanceFromCenter = hypot(location.x, location.y)/colorWheelRadius;
+	CGFloat relativeDistanceFromCenter = hypot(location.x, location.y)/colorWheelRadius;
 	if (relativeDistanceFromCenter > 1)
 		relativeDistanceFromCenter = 1;
 	
 	// Get the angle and ensure it's in 0..2*PI
-	float angle = fmod(atan2(location.y, location.x) - M_PI_2 + M_PI * 2.f, M_PI * 2.f);
+	CGFloat angle = fmod(atan2(location.y, location.x) - M_PI_2 + M_PI * 2.f, M_PI * 2.f);
 	
 	// Hue is stored in degrees
 	_hueValue = angle * 180.f / M_PI;
