+++ AppKit/NSImage.m
@@ -486,9 +486,9 @@ NSImageName const NSImageNameApplicationIcon = @"NSApplicationIcon";
      else {
       NSSize checkSize=[check size];
       NSSize bestSize=[best size];
-      float  checkArea=checkSize.width*checkSize.height;
-      float  bestArea=bestSize.width*bestSize.height;
-      float  desiredArea=size.width*size.height;
+      CGFloat  checkArea=checkSize.width*checkSize.height;
+      CGFloat  bestArea=bestSize.width*bestSize.height;
+      CGFloat  desiredArea=size.width*size.height;
       
       // downsampling is better than upsampling
       if(bestArea<desiredArea && checkArea>=desiredArea)
@@ -729,7 +729,7 @@ NSImageName const NSImageNameApplicationIcon = @"NSApplicationIcon";
    [self compositeToPoint:point fromRect:rect operation:operation fraction:1.0];
 }
 
--(void)compositeToPoint:(NSPoint)point fromRect:(NSRect)source operation:(NSCompositingOperation)operation fraction:(float)fraction {
+-(void)compositeToPoint:(NSPoint)point fromRect:(NSRect)source operation:(NSCompositingOperation)operation fraction:(CGFloat)fraction {
    /* Compositing is a blitting operation. We simulate it using the draw operation.
    
       Compositing does not honor all aspects of the CTM, e.g. it will keep an image upright regardless of the orientation of CTM.
@@ -756,25 +756,25 @@ NSImageName const NSImageNameApplicationIcon = @"NSApplicationIcon";
    [self compositeToPoint:point operation:operation fraction:1.0];
 }
 
--(void)compositeToPoint:(NSPoint)point operation:(NSCompositingOperation)operation fraction:(float)fraction {   
+-(void)compositeToPoint:(NSPoint)point operation:(NSCompositingOperation)operation fraction:(CGFloat)fraction {   
    [self compositeToPoint:point fromRect:NSZeroRect operation:operation fraction:1.0];
 }
 
--(void)dissolveToPoint:(NSPoint)point fraction:(float)fraction {
+-(void)dissolveToPoint:(NSPoint)point fraction:(CGFloat)fraction {
    NSUnimplementedMethod();
 }
 
--(void)dissolveToPoint:(NSPoint)point fromRect:(NSRect)rect fraction:(float)fraction {
+-(void)dissolveToPoint:(NSPoint)point fromRect:(NSRect)rect fraction:(CGFloat)fraction {
    NSUnimplementedMethod();
 }
 
--(void)drawAtPoint:(NSPoint)point fromRect:(NSRect)source operation:(NSCompositingOperation)operation fraction:(float)fraction {
+-(void)drawAtPoint:(NSPoint)point fromRect:(NSRect)source operation:(NSCompositingOperation)operation fraction:(CGFloat)fraction {
    NSSize size=[self size];
    
    [self drawInRect:NSMakeRect(point.x,point.y,size.width,size.height) fromRect:source operation:operation fraction:fraction];
 }
 
--(void)drawInRect:(NSRect)rect fromRect:(NSRect)source operation:(NSCompositingOperation)operation fraction:(float)fraction {
+-(void)drawInRect:(NSRect)rect fromRect:(NSRect)source operation:(NSCompositingOperation)operation fraction:(CGFloat)fraction {
 
 	// Keep a lid on any intermediate allocations while producing caches
 	NSAutoreleasePool *pool=[NSAutoreleasePool new];
