+++ AppKit/NSInterfacePart/NSInterfacePartAttributedString.h
@@ -15,7 +15,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
     NSAttributedString *_attributedString;
 }
 
-- initWithCharacter:(unichar)character fontName:(NSString *)fontName pointSize:(float)pointSize color:(NSColor *)color;
+- initWithCharacter:(unichar)character fontName:(NSString *)fontName pointSize:(CGFloat)pointSize color:(NSColor *)color;
 
 - initWithMarlettCharacter:(unichar)character;
