+++ AppKit/NSMenu.subproj/NSMainMenuView.h
@@ -12,7 +12,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
     NSFont *_font;
 }
 
-+ (float)menuHeight;
++ (CGFloat)menuHeight;
 
 - initWithFrame:(NSRect)frame menu:(NSMenu *)menu;
